:white_check_mark: Address review comments

+++ tiatoolbox/models/engine/semantic_segmentor.py
@@ -1369,7 +1369,7 @@ def prepare_full_batch(
     full_output_dict = {tuple(row): i for i, row in enumerate(full_output_locs)}
     matches = [full_output_dict[tuple(row)] for row in batch_locs]
 
-    total_size = np.max(matches).astype(np.uint16) + 1
+    total_size = np.max(matches).astype(np.uint32) + 1
 
     # Initialize full output array
     full_batch_output = np.zeros(
